fix(pay): move Alternate Oscar History aggregate computation server-side (PAY-2)
AlternateOscarHistoryPanel called useAlternateOscarHistory(movies) for
every user regardless of entitlement, then only CSS-blurred the real DOM
output for free users -- the actual lifetime upheld rate, by-decade
breakdown, and "most controversial calls" (specific movie titles and
ratings) were present in the DOM and React state the whole time,
recoverable via devtools or by disabling the blur class. Same class of
bug as PAY-4 from the 8/15 audit, still unpatched.

New GET /api/alternate-oscar-history checks isPremiumUser() and only
returns the real rollup to a premium account; a non-premium caller gets
`{ locked: true, hasEligibleYears }` -- enough to know whether to show a
locked teaser at all, nothing else. The panel now fetches this route
instead of computing from a `movies` prop, and renders neutral skeleton
placeholders (not blurred real data) under the lock overlay.

To make the same rollup logic safely callable from a Route Handler,
split three pure, Supabase-client-free modules out of files that
previously mixed pure logic with the browser-only Supabase client
(src/lib/supabaseBrowser.ts is explicitly documented as unsafe to import
into server code):
  - src/utils/normalizeUserAward.ts — toNormalizedAward, shared by
    useUserAwards.ts and the new route instead of duplicated.
  - src/utils/academyStatus.ts — getAcademyStatus/getAcademyContextMessage,
    split out of src/data/officialAwardWinners.ts (which re-exports them
    for existing call sites).
  - src/utils/alternateOscarHistory.ts — computeAlternateOscarHistory,
    split out of the useAlternateOscarHistory hook (which is now a thin
    wrapper around it, kept for any other already-premium client caller
    but no longer used by this panel).
